Moved all components in a folder components
