+++ src/components/TextMode.js
