+++ src/components/menu/entries.js
